Move AppShellExtensionHandler and StContextScope out of client_app
