+++ appshell_paths.gypi
@@ -147,6 +147,7 @@
     'appshell_sources_common_helper': [
       'appshell/common/client_switches.cc',
       'appshell/common/client_switches.h',
+      'appshell/appshell_extension_handler.h',
       'appshell/appshell_extensions.cpp',
       'appshell/appshell_extensions.h',
       'appshell/appshell_extensions_platform.h',
